pyproject.toml
[project]
name = "index-wrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "patool>=3.1.0",
    "rclone-python>=0.1.18",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "ruff>=0.8.4",
    "selenium>=4.27.1",
    "supabase>=2.10.0",
    "webdriver-manager>=4.0.2",
]